add change_copies_number Postgres function
changing copies number is tricky: when you just want to add it's easy, but when you want to deleting, I need to be sure that you aren't deleting copies in loan/consultation/reservation, so the copies that are effectively deleted are just the ones that are not in those categories. For the other ones, the DB will set 'reserved' as status so none can reserve them another time. Plus, the renew, if before was possible, now is no longer possible on loans that have that copies