bugfixes and tests improvements
add the field `done` to ConsultationReservation to find out if it's done
or not. The field was already in the db, but it wasn't an attribute in
the class. There was a method
`done() -> return true if this.done.isAfter(today)`, now it's properly handled
by the attribute.

All methods of `DbManager` have a test that runs them.