modify `getLoanReservationByUser`
The new method in `DbManager` uses two more params `useDoneParam` and
`doneParam`. These two params are used to retrieve the `LoanReservation`
that has *NOT* been  done yet.

First steps in modifying how the Librarian interacts when a user
starts a new reserved loan, because I want to follow the same pattern
of new reserved consultation.