improved study-based flow
change `addStudy` to `tryAddStudy` using a PG function: we insert the
study only if it's not already there. This is very important because
when the librarian calls `start*ReservedConsultation` a new study is
added, but a new study is added also when turnstile calls `getOrAssignSeat`,
and there's no way (not an easy one) to check if the seat returnned comes from
a study (consultation) or has been assigned now (study not reserved) or
previously (study reserved).

add table to view the seats situtation, only available to librarian.